pyproject.toml
[project]
name = "skill-dev-sessions"
version = "0.1.0"
description = ""
authors = [
    {name = "Parul Chutani",email = "parul.chutani@thepsi.com"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.10,<0.16.0)",
    "black (>=26.3.1,<27.0.0)"
]
